Running data sync via cron
